face/requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.4
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-crispy-forms==1.11.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
idna==2.10
importlib-metadata==3.4.0
kombu==5.0.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.1.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytube==10.4.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
